feat(#37): added comment on PR when no tests are included
Before it adds a comment it checks all present comments in the issue/pull-request. If no comment with main plugins title "Ike Plugins" is found, then it adds a new comment with the main title, assignee mention, the test-keeper plugin tittle and the plugin commentMsg. If a comment with the main title is found but missing the test-keeper plugin title, then it modifies the comment by adding the plugin title with the comment message. If everything is present already, then it does nothing.

Comment message is constructed based on the fact if configuration is
present. If config file is not present, then it shows default comment
referencing to project documentation. If config is present but missing custom
comment file there then it shows default with reference to the config
file. If config file with custom comment file is present, then it
shows the given custom message.